Added dock state receiver and events
I added a receiver for the dock state and dock/undock events that get
populated on the server. Doesn't restart IO if undocked (it should
probably handle that better). Currently the actual dock state is passed
as additional data but in the end it should be added to the standard
message structure, at least for OBC5/MT5.

NOTE: this changes aren't compatible with the A317 now so additional
code will have to be added to make sure the A317 still builds and works
properly.