Added encoded dock state
The dock state is sent in the messages to the server as the last byte.
If there is no additional data then 00 00 is put so that we can put the
dock state after the additional data. There is a still an issue where
ignition changes before dock state changes, so on ignition changes the
dock state sent with the event is often wrong.